Tests
Encoding gives a float and the difference (a float) which, added to
the decoded value, returns the original float. Currently all tests
passes, which means the difference is well computed. Also, the
difference seems always small. However, the new tests
when *check-nearest-encoding* is T fail for some values, which means
there are floats for which neighbour BITS (1+ BITS) or (1- BITS) are
more accurate encodings.